+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Improved directory cache behavior
+
 ## [0.5.1] - 2026-04-24
 
 ### Added